Move webhook handling into api4 package to fix EnableAPIv3 config setting
